finishing user and roles

+++ routes/web.php
@@ -15,8 +15,9 @@ use App\Http\Controllers\MembershipController;
 use App\Http\Controllers\LogMembershipController;
 
 Route::get('/', [LandingController::class, 'index'])->name('home');
-Route::get('/register-new-member', [LandingController::class, 'store'])->name('home.store');
+Route::post('/register-new-member', [LandingController::class, 'store'])->name('home.store');
 Route::get('/registration-success', [LandingController::class, 'success'])->name('home.success');
+Route::get('/gym-package', [LandingController::class, 'gym_package'])->name('home.gym_package');
 
 
 Route::middleware(['auth'])->group(function () {
@@ -34,7 +35,7 @@ Route::middleware(['auth'])->group(function () {
             'create' => 'membership.create',
             'edit' => 'membership.edit',
             'destroy' => 'membership.destroy',
-        ]);
+        ])->middleware('role:super_admin|admin');
     Route::post('membership/reset-password/{id}', [MembershipController::class, 'reset_password'])->name('membership.reset-password');
 
     Route::resource('extend-membership', LogMembershipController::class)
@@ -48,26 +49,26 @@ Route::middleware(['auth'])->group(function () {
             'create' => 'extend-membership.create',
             'store' => 'extend-membership.store',
         ]);
-    Route::post('extend-membership/proses_approve_reject/{id}/{type}', [LogMembershipController::class, 'proses_approve_reject'])->name('membership.proses_approve_reject');
+    Route::post('extend-membership/proses_approve_reject/{id}/{type}', [LogMembershipController::class, 'proses_approve_reject'])->name('membership.proses_approve_reject')->middleware('role:super_admin|admin');
 
-    Route::get('manage-admin', [ManageAdminController::class, 'index'])->name('manage-admin');
-    Route::get('manage-admin/create', [ManageAdminController::class, 'create'])->name('manage-admin.create');
-    Route::get('manage-admin/edit/{user}', [ManageAdminController::class, 'edit'])->name('manage-admin.edit');
-    Route::delete('manage-admin/destroy/{user}', [ManageAdminController::class, 'destroy'])->name('manage-admin.destroy');
-    Route::post('manage-admin/reset-password/{id}', [ManageAdminController::class, 'reset_password'])->name('manage-admin.reset-password');
+    Route::get('manage-admin', [ManageAdminController::class, 'index'])->name('manage-admin')->middleware('role:super_admin');
+    Route::get('manage-admin/create', [ManageAdminController::class, 'create'])->name('manage-admin.create')->middleware('role:super_admin');
+    Route::get('manage-admin/edit/{user}', [ManageAdminController::class, 'edit'])->name('manage-admin.edit')->middleware('role:super_admin');
+    Route::delete('manage-admin/destroy/{user}', [ManageAdminController::class, 'destroy'])->name('manage-admin.destroy')->middleware('role:super_admin');
+    Route::post('manage-admin/reset-password/{id}', [ManageAdminController::class, 'reset_password'])->name('manage-admin.reset-password')->middleware('role:super_admin');
 
-    Route::get('manage-paket', [GymPackageController::class, 'index'])->name('manage-paket');
-    Route::get('manage-paket/create', [GymPackageController::class, 'create'])->name('manage-paket.create');
-    Route::get('manage-paket/edit/{gym_package}', [GymPackageController::class, 'edit'])->name('manage-paket.edit');
-    Route::delete('manage-paket/destroy/{gym_package}', [GymPackageController::class, 'destroy'])->name('manage-paket.destroy');
+    Route::get('manage-paket', [GymPackageController::class, 'index'])->name('manage-paket')->middleware('role:super_admin|admin');
+    Route::get('manage-paket/create', [GymPackageController::class, 'create'])->name('manage-paket.create')->middleware('role:super_admin|admin');
+    Route::get('manage-paket/edit/{gym_package}', [GymPackageController::class, 'edit'])->name('manage-paket.edit')->middleware('role:super_admin|admin');
+    Route::delete('manage-paket/destroy/{gym_package}', [GymPackageController::class, 'destroy'])->name('manage-paket.destroy')->middleware('role:super_admin|admin');
 
-    Route::get('manage-carousel', [CarouselController::class, 'index'])->name('manage-carousel');
-    Route::get('manage-carousel/create', [CarouselController::class, 'create'])->name('manage-carousel.create');
-    Route::get('manage-carousel/edit/{carousel}', [CarouselController::class, 'edit'])->name('manage-carousel.edit');
-    Route::delete('manage-carousel/destroy/{carousel}', [CarouselController::class, 'destroy'])->name('manage-carousel.destroy');
+    Route::get('manage-carousel', [CarouselController::class, 'index'])->name('manage-carousel')->middleware('role:super_admin|admin');
+    Route::get('manage-carousel/create', [CarouselController::class, 'create'])->name('manage-carousel.create')->middleware('role:super_admin|admin');
+    Route::get('manage-carousel/edit/{carousel}', [CarouselController::class, 'edit'])->name('manage-carousel.edit')->middleware('role:super_admin|admin');
+    Route::delete('manage-carousel/destroy/{carousel}', [CarouselController::class, 'destroy'])->name('manage-carousel.destroy')->middleware('role:super_admin|admin');
 
-    Route::get('info-gym', [InfoGymController::class, 'index'])->name('info-gym');
-    Route::post('info-gym/update', [InfoGymController::class, 'update'])->name('info-gym.update');
+    Route::get('info-gym', [InfoGymController::class, 'index'])->name('info-gym')->middleware('role:super_admin|admin');
+    Route::post('info-gym/update', [InfoGymController::class, 'update'])->name('info-gym.update')->middleware('role:super_admin|admin');
 
     Route::redirect('settings', 'settings/profile');
